+++ Demo/Application/Features/IdealViewController.swift
@@ -1,8 +1,8 @@
-import Foundation
+import UIKit
 import BraintreeLocalPayment
 import BraintreeCore
 
-class BraintreeDemoIdealViewController: BraintreeDemoPaymentButtonBaseViewController {
+class IdealViewController: PaymentButtonBaseViewController {
 
     var localPaymentClient: BTLocalPaymentClient!
     var paymentIDLabel: UILabel = UILabel()
@@ -11,20 +11,12 @@ class BraintreeDemoIdealViewController: BraintreeDemoPaymentButtonBaseViewContro
         super.viewDidLoad()
 
         progressBlock("Loading iDEAL Merchant Account...")
-        paymentButton.isHidden = false
         progressBlock("Ready!")
         title = "iDEAL"
     }
 
-    override func createPaymentButton() -> UIView! {
-        let iDEALButton = UIButton(type: .custom)
-        iDEALButton.setTitle("Pay with iDEAL", for: .normal)
-        iDEALButton.setTitleColor(.blue, for: .normal)
-        iDEALButton.setTitleColor(.lightGray, for: .highlighted)
-        iDEALButton.setTitleColor(.lightGray, for: .disabled)
-        iDEALButton.addTarget(self, action: #selector(tappedIDEAL), for: .touchUpInside)
-        iDEALButton.translatesAutoresizingMaskIntoConstraints = false
-
+    override func createPaymentButton() -> UIView {
+        let iDEALButton = createButton(title: "Pay with iDEAL", action: #selector(tappedIDEAL))
         let label = UILabel()
         label.numberOfLines = 0
         label.translatesAutoresizingMaskIntoConstraints = false
@@ -95,7 +87,7 @@ class BraintreeDemoIdealViewController: BraintreeDemoPaymentButtonBaseViewContro
 
 // MARK: - BTLocalPaymentRequestDelegate Conformance
 
-extension BraintreeDemoIdealViewController: BTLocalPaymentRequestDelegate {
+extension IdealViewController: BTLocalPaymentRequestDelegate {
 
     func localPaymentStarted(
         _ request: BraintreeLocalPayment.BTLocalPaymentRequest,