+++ Demo/Application/Features/PayPalWebCheckoutViewController.swift
@@ -1,0 +1,81 @@
+import Foundation
+import UIKit
+import BraintreePayPal
+
+class PayPalWebCheckoutViewController: PaymentButtonBaseViewController {
+
+    lazy var payPalClient = BTPayPalClient(apiClient: apiClient)
+
+    override func createPaymentButton() -> UIView {
+        let payPalCheckoutButton = createButton(title: "PayPal Checkout", action: #selector(tappedPayPalCheckout))
+        let payPalVaultButton = createButton(title: "PayPal Vault", action: #selector(tappedPayPalVault))
+        let payPalPayLaterButton = createButton(title: "PayPal with Pay Later Offered", action: #selector(tappedPayPalPayLater))
+
+        let buttons = [payPalCheckoutButton, payPalVaultButton, payPalPayLaterButton]
+        let stackView = UIStackView(arrangedSubviews: buttons)
+        stackView.axis = .vertical
+        stackView.alignment = .center
+        stackView.distribution = .fillEqually
+        stackView.translatesAutoresizingMaskIntoConstraints = false
+        
+        return stackView
+    }
+
+    @objc func tappedPayPalCheckout(_ sender: UIButton) {
+        progressBlock("Tapped PayPal - Checkout using BTPayPalClient")
+        sender.setTitle("Processing...", for: .disabled)
+        sender.isEnabled = false
+
+        let request = BTPayPalCheckoutRequest(amount: "4.30")
+
+        payPalClient.tokenize(request) { nonce, error in
+            sender.isEnabled = true
+
+            guard let nonce else {
+                self.progressBlock(error?.localizedDescription)
+                return
+            }
+
+            self.completionBlock(nonce)
+        }
+    }
+
+    @objc func tappedPayPalVault(_ sender: UIButton) {
+        progressBlock("Tapped PayPal - Vault using BTPayPalClient")
+        sender.setTitle("Processing...", for: .disabled)
+        sender.isEnabled = false
+
+        let request = BTPayPalVaultRequest()
+
+        payPalClient.tokenize(request) { nonce, error in
+            sender.isEnabled = true
+
+            guard let nonce else {
+                self.progressBlock(error?.localizedDescription)
+                return
+            }
+
+            self.completionBlock(nonce)
+        }
+    }
+
+    @objc func tappedPayPalPayLater(_ sender: UIButton) {
+        progressBlock("Tapped PayPal - initiating with Pay Later offered")
+        sender.setTitle("Processing...", for: .disabled)
+        sender.isEnabled = false
+
+        let request = BTPayPalCheckoutRequest(amount: "4.30")
+        request.offerPayLater = true
+
+        payPalClient.tokenize(request) { nonce, error in
+            sender.isEnabled = true
+
+            guard let nonce else {
+                self.progressBlock(error?.localizedDescription)
+                return
+            }
+
+            self.completionBlock(nonce)
+        }
+    }
+}