+++ Demo/Application/Supporting Files/Demo-Bridging-Header.h
@@ -1,3 +1,0 @@
-#import "BraintreeDemoBaseViewController.h"
-#import <BraintreeVenmo/BraintreeVenmo-Swift.h>
-#import "BraintreeDemoPaymentButtonBaseViewController.h"