+++ Sources/BraintreeCore/BTAPIClient.swift
@@ -203,15 +203,20 @@ import Foundation
     /// Fetches a customer's vaulted payment method nonces.
     /// Must be using client token with a customer ID specified.
     ///  - Parameter completion: Callback that returns either an array of payment method nonces or an error
+    ///  - Note: Only the top level `BTPaymentMethodNonce` type is returned, fetching any additional details will need to be done on the server
     public func fetchPaymentMethodNonces(_ completion: @escaping ([BTPaymentMethodNonce]?, Error?) -> Void) {
         fetchPaymentMethodNonces(false, completion: completion)
     }
 
+    // NEXT_MAJOR_VERSION: this should move into the Drop-in for parity with Android
+    // This will also allow us to return the types directly which we were doing in the +load method
+    // previously in Obj-C - this is not available in Swift
     /// Fetches a customer's vaulted payment method nonces.
     /// Must be using client token with a customer ID specified.
     ///  - Parameters:
     ///   - defaultFirst: Specifies whether to sort the fetched payment method nonces with the default payment method or the most recently used payment method first
     ///   - completion: Callback that returns either an array of payment method nonces or an error
+    ///   - Note: Only the top level `BTPaymentMethodNonce` type is returned, fetching any additional details will need to be done on the server
     public func fetchPaymentMethodNonces(_ defaultFirst: Bool, completion: @escaping ([BTPaymentMethodNonce]?, Error?) -> Void) {
         if clientToken == nil {
             completion(nil, BTAPIClientError.notAuthorized)
@@ -318,10 +323,11 @@ import Foundation
 
     /// :nodoc: This method is exposed for internal Braintree use only. Do not use. It is not covered by Semantic Versioning and may change or be removed at any time.
     @_documentation(visibility: private)
-    public func sendAnalyticsEvent(_ eventName: String, errorDescription: String? = nil) {
+    public func sendAnalyticsEvent(_ eventName: String, errorDescription: String? = nil, correlationID: String? = nil) {
         analyticsService?.sendAnalyticsEvent(
             eventName,
             errorDescription: errorDescription,
+            correlationID: correlationID,
             completion: { _ in }
         )
     }