+++ UnitTests/BraintreeCoreTests/Analytics/FPTIBatchData_Tests.swift
@@ -17,11 +17,13 @@ final class FPTIBatchData_Tests: XCTestCase {
     
     let eventParams = [
         FPTIBatchData.Event(
+            correlationID: "fake-correlation-id-1",
             errorDescription: "fake-error-description-1",
             eventName: "fake-event-1",
             timestamp: "fake-time-1"
         ),
         FPTIBatchData.Event(
+            correlationID: nil,
             errorDescription: nil,
             eventName: "fake-event-2",
             timestamp: "fake-time-2"
@@ -58,7 +60,7 @@ final class FPTIBatchData_Tests: XCTestCase {
         XCTAssertEqual(batchParams["auth_fingerprint"] as! String, "fake-auth")
         XCTAssertTrue((batchParams["c_sdk_ver"] as! String).matches("^\\d+\\.\\d+\\.\\d+(-[0-9a-zA-Z-]+)?$"))
         XCTAssertTrue((batchParams["client_os"] as! String).matches("iOS \\d+\\.\\d+|iPadOS \\d+\\.\\d+"))
-        XCTAssertEqual(batchParams["comp"] as? String, "btmobilesdk")
+        XCTAssertEqual(batchParams["comp"] as? String, "braintreeclientsdk")
         XCTAssertEqual(batchParams["device_manufacturer"] as? String, "Apple")
         XCTAssertEqual(batchParams["merchant_sdk_env"] as? String, "fake-env")
         XCTAssertEqual(batchParams["event_source"] as? String, "mobile-native")
@@ -70,7 +72,6 @@ final class FPTIBatchData_Tests: XCTestCase {
         XCTAssertEqual(batchParams["merchant_id"] as! String, "fake-merchant-id")
         XCTAssertEqual(batchParams["platform"] as? String, "iOS")
         XCTAssertEqual(batchParams["session_id"] as? String, "fake-session")
-        XCTAssertEqual(batchParams["tenant_name"] as? String, "Braintree")
         XCTAssertEqual(batchParams["tokenization_key"] as! String, "fake-auth")
 
         // Verify event-level parameters
@@ -78,7 +79,11 @@ final class FPTIBatchData_Tests: XCTestCase {
         XCTAssertEqual(eventParams[1]["t"] as? String, "fake-time-2")
         XCTAssertEqual(eventParams[0]["event_name"] as? String, "fake-event-1")
         XCTAssertEqual(eventParams[1]["event_name"] as? String, "fake-event-2")
+        XCTAssertEqual(eventParams[0]["tenant_name"] as? String, "Braintree")
+        XCTAssertEqual(eventParams[1]["tenant_name"] as? String, "Braintree")
         XCTAssertEqual(eventParams[0]["error_desc"] as? String, "fake-error-description-1")
         XCTAssertNil(eventParams[1]["error_desc"])
+        XCTAssertEqual(eventParams[0]["correlation_id"] as? String, "fake-correlation-id-1")
+        XCTAssertNil(eventParams[1]["correlation_id"])
     }
 }