+++ UnitTests/BraintreeCoreTests/BTAPIClient_Tests.swift
@@ -248,9 +248,11 @@ class BTAPIClient_Tests: XCTestCase {
 
             let firstNonce = paymentMethodNonces[0];
             XCTAssertEqual(firstNonce.nonce, "fake-nonce1")
+            XCTAssertEqual(firstNonce.type, "AMEX")
 
             let secondNonce = paymentMethodNonces[1]
             XCTAssertEqual(secondNonce.nonce, "fake-nonce2")
+            XCTAssertEqual(secondNonce.type, "PayPal")
 
             expectation.fulfill()
         }